BACKEND/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.3
httpx==0.28.1
beautifulsoup4==4.13.4
python-dotenv==1.1.0
sentence-transformers==4.1.0
chromadb==0.6.3
rank-bm25==0.2.2
slowapi==0.1.9
